+++ src/routes/EcosystemDeposit/MellowLpDepositForm/components/LPMellowVaultDepositInfo.styled.tsx
